requirements.txt
boto3==1.34.46
botocore==1.34.46
certifi==2024.2.2
charset-normalizer==3.3.2
greenlet==3.0.3
idna==3.6
jmespath==1.0.1
load-dotenv==0.1.0
numpy==1.26.4
pandas==2.2.0
Pandas3==0.0.1
pprint36==3.9.16.0
psycopg2==2.9.9
psycopg2-binary==2.9.9
pyarrow==15.0.0
PyMuPDF==1.23.25
PyMuPDFb==1.23.22
python-dateutil==2.8.2
python-dotenv==1.0.1
pytz==2024.1
requests==2.31.0
s3transfer==0.10.0
six==1.16.0
SQLAlchemy==2.0.27
typing_extensions==4.9.0
tzdata==2024.1
urllib3==2.0.7
